requirements.txt
psycopg2-binary>=2.9.9
openpyxl>=3.1.2
python-dotenv>=1.0.0
tqdm>=4.66.1
flask>=2.0.1
python-dateutil>=2.8.2
